Fixed AllGlasses Component. Moved to components folder
